requirements.txt
Flask
tensorflow
pillow
numpy
mysql-connector-python
python-dotenv
gunicorn